pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-i18n"
version = "0.3.9"
description = "Translation library for Python"
readme = "README.md"
authors = [
    { name = "Daniel Perez", email = "tuvistavie@gmail.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.7"
dependencies = [
    "pyyaml>=3.10"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries"
]

[project.optional-dependencies]
yaml = ["pyyaml>=3.10"]

[tool.setuptools.packages.find]
where = ["i18n"]
include = ["i18n*", "i18n.loaders*", "i18n.tests*"]
